style : modify tourist Spot card style
