+++ src/mocks/get/home.js
@@ -20,43 +20,3 @@ export const getHomeHandler = rest.get("/home", (req, res, ctx) => {
     }),
   );
 });
-
-// export const searchSpotHandler = rest.get("/api/search", (req, res, ctx) => {
-//   const query = req.url.searchParams.get("query");
-
-//   if (query) {
-//     return res(
-//       ctx.status(200),
-//       ctx.json({
-//         success: true,
-//         response: {
-//           restaurants: getRestaurants(8),
-//           festivals: getFestivals(8),
-//           touristSpots: getTouristSpots(8),
-//         },
-//       }),
-//     );
-//   }
-
-//   return res(ctx.status(200), ctx.json({ result: {} }));
-// });
-
-export const getSearchHandler = rest.get("/search", (req, res, ctx) => {
-  const query = req.url.searchParams.get("query");
-
-  if (query) {
-    return res(
-      ctx.status(200),
-      ctx.json({
-        success: true,
-        response: {
-          restaurants: getRestaurants(8),
-          festivals: getFestivals(8),
-          touristSpots: getTouristSpots(8),
-        },
-      }),
-    );
-  }
-
-  return res(ctx.status(200), ctx.json({ result: {} }));
-});
